Move RosterMetrics
