+++ platform-sdk/consensus-roster/src/main/java/org/hiero/consensus/roster/RosterMetrics.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.state.address;
+package org.hiero.consensus.roster;
 
 import com.hedera.hapi.node.state.roster.Roster;
 import com.swirlds.metrics.api.Metrics;
